Move infer functionality into ccj module
Consolidate compile_commands.json related functionality.

* include/blot/ccj.hpp: Add infer function declaration.

* src/libblot/ccj.cpp: Move infer and create_translation_unit functions.
Add required includes for libclang functionality.

* test/infer-tests.cpp: Include ccj.hpp instead of infer.hpp.

* include/blot/infer.hpp: Remove file.

* src/libblot/infer.cpp: Remove file.
